Add buildkite and earthfile configs for CI
This adds an earthfile for doing the actual building of the site, and
then a buildkite file which will run that build and then handle pushing
things to S3 and clearing the Cloudfront cache.

One change to how we get the `protocol.json` file: we cannot always
expect to have the `backend` directory as a sibling folder in CI, so
instead we clone down the latest version of the protocol package (which
is what we should be building the docs for anyways), and use the
`protocol.json` from our `node_modules` directory.